requirements.txt
kivymd
kivy
plyer
pillow
google-generativeai
requests
certifi
psycopg2-binary
python-dotenv
